--- a/axum/Cargo.toml
+++ b/axum/Cargo.toml
@@ -95,4 +95,5 @@ features = [
     "json",
     "multipart",
     "ws",
+    "utf8_unckecked",
 ]