Like this comment:

```rust
// `ByteStr` can only be constructed from strings which are always valid
// utf-8 so this wont panic.
std::str::from_utf8(&self.0).unwrap()
```

We know it always valid, But axum promise `This crate uses #![forbid(unsafe_code)] to ensure everything is implemented in 100% safe Rust.` 

So, i think we can add a feature that user can use `unsafe` by themself.