@jplatte 2022-04-24
This makes some sense to me and if it shows up on benchmarks I'd be in favor, but *not* if it's behind a feature flag like this.

Personally I think `forbid(unsafe_code)` is overrated, but @davidpdrsn will probably disagree, given he added it to this project 😄