@NateLing 2022-04-24
@jplatte  I know this PR breaks Axum's promises, but I think default promises and all promises are different promises. For different consumers, they are free to choose.