@davidpdrsn 2022-04-24
This particular use of unsafe is totally fine but yeah I still think we should keep `#[forbid(unsafe)]` unless we have a good reason to remove it. So unless this measurably improves performance then I don't think we should change anything.

Regardless it shouldn't be an optional feature, that feels kinda odd to me 🤷