@NateLing 2022-04-24
@davidpdrsn I‘ll add a benchmark, maybe tomorrow