@davidpdrsn 2022-04-24
You don't need to add a benchmark to this repo. You can run one of the benchmarks mentioned [here](https://github.com/tokio-rs/axum#performance) and report your results.