@davidpdrsn 2022-04-25
I've benchmarked this now and don't see any difference.

Code:

```rust
use axum::{extract::Path, response::IntoResponse, routing::get, Router};
use std::net::SocketAddr;

#[tokio::main]
async fn main() {
    let app = Router::new().route("/:a/:b/:c/:d/:e", get(handler));

    let addr = SocketAddr::from(([127, 0, 0, 1], 3000));
    axum::Server::bind(&addr)
        .serve(app.into_make_service())
        .await
        .unwrap();
}

#[derive(serde::Deserialize)]
struct Params {
    a: String,
    b: String,
    c: String,
    d: String,
    e: String,
}

async fn handler(_: Path<Params>) -> impl IntoResponse {}
```

Command (each path param is a uuid):

```
wrk \
    -t12 \
    -c400 \
    -d30s \
    http://127.0.0.1:3000/cd0e50d0-db9a-481a-9499-cd0e50d0-db9a-481a-9499-b8179a97a35c/cd0e50d0-db9a-481a-9499-b8179a97a35c/cd0e50d0-db9a-481a-9499-b8179a97a35c/cd0e50d0-db9a-481a-9499-b8179a97a35c/cd0e50d0-db9a-481a-9499-b8179a97a35c
```

Baseline (without unsafe):

```
  12 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.56ms    4.51ms 252.66ms   99.08%
    Req/Sec    15.74k     7.17k   36.37k    56.46%
  5641965 requests in 30.10s, 403.54MB read
  Socket errors: connect 155, read 42, write 0, timeout 0
Requests/sec: 187420.96
Transfer/sec:     13.41MB
```

With unsafe:

```
  12 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.66ms    5.36ms 251.46ms   99.10%
    Req/Sec    15.66k     8.25k   39.13k    53.27%
  5602355 requests in 30.09s, 400.71MB read
  Socket errors: connect 155, read 103, write 0, timeout 0
Requests/sec: 186170.38
Transfer/sec:     13.32MB
```

So actually a small decrease in performance. 187420.96 requests/second to 186170.38 requests/second. That could just be noise but it still isn't an improvement.

@NateLing have you done your own tests?