@NateLing 2022-04-25
@davidpdrsn  Not yet, I am busy in this [Add rest and grpc example](https://github.com/tokio-rs/axum/pull/967)