@davidpdrsn 2022-04-25
I tested with a handle that used no extractors and that was significantly faster, so I don't think so.