@NateLing 2022-04-26
@davidpdrsn I've benchmarked and got a strange result

code:
```rust
use axum::{extract::Path, routing::get, Router};
use std::net::SocketAddr;

#[tokio::main]
async fn main() {
    let app = Router::new().route("/:a/:b/:c/:d/:e", get(handler));

    let addr = SocketAddr::from(([127, 0, 0, 1], 3000));
    axum::Server::bind(&addr)
        .serve(app.into_make_service())
        .await
        .unwrap();
}

#[derive(serde::Deserialize)]
struct Params {
    a: String,
    b: String,
    c: String,
    d: String,
    e: String,
}

async fn handler(path: Path<Params>) -> String {
    let Path(Params { a, b, c, d, e }) = path;
    [a, b, c, d, e].join("/")
}

```

Profile release

```toml
[profile.release]
strip = false
opt-level = 3
debug = false
debug-assertions = false
overflow-checks =false
lto = "fat"
incremental = false
codegen-units = 1
panic = "abort"
rpath = false
```

rustc --version --verbose

```bash
rustc 1.62.0-nightly (055bf4ccd 2022-04-25)
binary: rustc
commit-hash: 055bf4ccd521c2c2185166c86951be7be145727c
commit-date: 2022-04-25
host: x86_64-apple-darwin
release: 1.62.0-nightly
LLVM version: 14.0.1
```

wrk :

```bash
wrk  -t4  -c100  -d30s  http://127.0.0.1:3000/8fc3a2ac-c504-11ec-9d64-0242ac120002/601da104-eed9-412b-b79d-df8e0dbc311e/008e1845-9be7-4881-b40c-1f16770115ad/0f1041a1-5224-491b-b67f-70fd569bdbdc/b6f2319c-50fa-4343-bb90-cb5ed261d7a8
```
Baseline (without unsafe):
```bash

  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.60ms    1.30ms  22.91ms   70.15%
    Req/Sec     6.15k   506.91    15.36k    77.14%
  735059 requests in 30.10s, 211.70MB read
Requests/sec:  24423.79
Transfer/sec:      7.03MB


  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.60ms    1.30ms  18.30ms   69.85%
    Req/Sec     6.13k   544.66    18.40k    81.64%
  732764 requests in 30.09s, 211.04MB read
Requests/sec:  24354.29
Transfer/sec:      7.01MB


  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.60ms    1.30ms  13.64ms   69.85%
    Req/Sec     6.15k   453.98    11.55k    73.53%
  735911 requests in 30.09s, 211.95MB read
Requests/sec:  24456.03
Transfer/sec:      7.04MB

```
With unsafe:

```bash

  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.72ms    1.34ms  34.41ms   69.89%
    Req/Sec     5.85k   368.87     7.53k    68.29%
  699168 requests in 30.02s, 201.37MB read
Requests/sec:  23290.32
Transfer/sec:      6.71MB


  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.71ms    1.35ms  23.13ms   69.44%
    Req/Sec     5.87k   560.64    20.43k    88.43%
  701511 requests in 30.10s, 202.04MB read
Requests/sec:  23309.25
Transfer/sec:      6.71MB


  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.75ms    1.49ms  37.67ms   74.44%
    Req/Sec     5.81k   444.13    12.69k    76.72%
  694208 requests in 30.09s, 199.94MB read
Requests/sec:  23069.42
Transfer/sec:      6.64MB

```

I really dont know why?