@NateLing 2022-04-26
Jemalloc   benchmarked result: 

code:
```rust
use axum::{extract::Path, routing::get, Router};
use std::net::SocketAddr;
use tikv_jemallocator::Jemalloc;

#[global_allocator]
static GLOBAL: Jemalloc = Jemalloc;

#[tokio::main]
async fn main() {
    let app = Router::new().route("/:a/:b/:c/:d/:e", get(handler));

    let addr = SocketAddr::from(([127, 0, 0, 1], 3000));
    axum::Server::bind(&addr)
        .serve(app.into_make_service())
        .await
        .unwrap();
}

#[derive(serde::Deserialize)]
struct Params {
    a: String,
    b: String,
    c: String,
    d: String,
    e: String,
}

async fn handler(path: Path<Params>) -> String {
    let Path(Params { a, b, c, d, e }) = path;
    [a, b, c, d, e].join("/")
}

```

Add `tikv-jemallocator` dependencies

```toml

[dependencies]
tikv-jemallocator = "0.4"

```



wrk :

```bash
wrk  -t4  -c100  -d30s  http://127.0.0.1:3000/8fc3a2ac-c504-11ec-9d64-0242ac120002/601da104-eed9-412b-b79d-df8e0dbc311e/008e1845-9be7-4881-b40c-1f16770115ad/0f1041a1-5224-491b-b67f-70fd569bdbdc/b6f2319c-50fa-4343-bb90-cb5ed261d7a8
```
Baseline (without unsafe):
```bash

  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.53ms    1.37ms  28.26ms   68.48%
    Req/Sec     4.02k   204.65     4.80k    69.19%
  480791 requests in 30.06s, 138.47MB read
Requests/sec:  15994.64
Transfer/sec:      4.61MB

  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.53ms    1.37ms  23.90ms   68.46%
    Req/Sec     4.03k   281.04    10.99k    85.03%
  481595 requests in 30.09s, 138.70MB read
Requests/sec:  16003.09
Transfer/sec:      4.61MB


  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.55ms    1.40ms  29.63ms   69.91%
    Req/Sec     4.01k   218.94     6.51k    74.24%
  478978 requests in 30.06s, 137.95MB read
Requests/sec:  15936.34
Transfer/sec:      4.59MB

```
With unsafe:

```bash

  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.47ms    1.32ms  23.95ms   68.46%
    Req/Sec     4.08k   258.17     9.93k    81.23%
  487489 requests in 30.07s, 140.40MB read
Requests/sec:  16212.90
Transfer/sec:      4.67MB

  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.48ms    1.34ms  20.60ms   68.22%
    Req/Sec     4.08k   300.41    11.92k    88.90%
  488227 requests in 30.10s, 140.61MB read
Requests/sec:  16222.25
Transfer/sec:      4.67MB


  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.47ms    1.32ms  21.23ms   68.55%
    Req/Sec     4.08k   208.42     6.64k    71.80%
  487960 requests in 30.05s, 140.54MB read
Requests/sec:  16236.35
Transfer/sec:      4.68MB

```

Tiny victory.